Rename util files for pattern matching
